fuck: use org-mode instead of Pandoc
Pandoc is turning out to be an uphill battle. There are a number of reasons for
not using it:
1. It is huge
2. Using it is very awkward because of very generic AST it produces
3. Its goals are very different. It isn't an org-mode parser, it is an org-mode
   exporter. It parses org-mode *after* catering to export settings defined in
   org preface, and then throws away a bunch of information. There is also no
   access to parser state, so working on links found in the org doc is going to
   be very difficult.

Because of these, I have decided to ditch Pandoc and create an org-mode parser
of my own. I've found https://github.com/fosskers/org-mode, which is a good
starting point. I intend to contribute back to it, but I was unable to figure
out how to setup proper dev-environment with nix flakes which don't make cabal
re-compile every single library. So in spirit of not breaking my momentum even
further, I've just copied org-mode over.

This commit is really just the starting point. `direnv reload` works, cabal can
build some files in `src/saunf`; but a lot of work is left. I have to redo the
whole readme functionality, rework the CLI. I think I am going to also need to
write an org-to-markdown compiler. A lot of rework ahead before I am on ground
zero to continue with saunf. Oh and now that we don't have Pandoc, I'll also
have to implement a whole lot of org-mode parsing myself. Not sure if it is a
good thing or bad; using pandoc was a learning experience, this will teach me
even more, but boy do it throw a wrench in my plan to get things done.