feat: get =saunf fmt= to work
Albeit incorrectly, but that's on org-mode. I need to write a proper
pretty-printer for org-mode myself. Which is sad. In the meanwhile though, I
think I can at least port all my code from Pandoc. Pandoc was a big help, I'll
have to do a lot of work just to make very little progress with org-mode